pyproject.toml
[project]
name = "git-whisperer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "rich>=13.7.0",
]
